# Conflicts:
#	src/main/java/co/edu/uceva/noticiasservice/controller/NoticiaRestController.java
#	src/main/java/co/edu/uceva/noticiasservice/model/service/NoticiaService.java
#	src/main/java/co/edu/uceva/noticiasservice/model/service/NoticiaServiceImpl.java

+++ src/main/java/co/edu/uceva/noticiasservice/controller/NoticiaRestController.java
@@ -4,9 +4,12 @@ package co.edu.uceva.noticiasservice.controller;
 import co.edu.uceva.noticiasservice.model.entities.Noticia;
 import co.edu.uceva.noticiasservice.model.service.NoticiaService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.http.HttpStatus;
+
+
+import java.util.List;
 
 import java.util.HashMap;
 import java.util.List;
@@ -20,7 +23,8 @@ public class NoticiaRestController {
 
     //crear noticia
     @PostMapping("/noticia")
-    public Noticia createNoticia(@RequestBody Noticia noticia) {
+    public Noticia createNoticia(@RequestBody Noticia noticia)
+    {
         return this.noticiaService.save(noticia);
     }
 
@@ -28,25 +32,23 @@ public class NoticiaRestController {
     @DeleteMapping("/noticia/{id}")
     public void deleteNoticia(@PathVariable int id) {
         Noticia noticia = this.noticiaService.findById(id);
-        this.noticiaService.delete(noticia);
+        if(noticia != null) {
+            this.noticiaService.delete(noticia);
+        }
     }
 
+    // Listar todas las noticias (para usuarios regulares, solo muestra las no eliminadas)
     @GetMapping("/noticia")
-    public ResponseEntity<?> getAllCelulars() {
-        Map<String, Object> response = new HashMap<>();
-        List<Noticia> noticia = null;
-        try {
-            noticia = this.noticiaService.listar();
-        } catch (Exception e) {
-            response.put("mensaje", "Error al listar los celulares");
-            response.put("error", e.getMessage());
-            return ResponseEntity.status(500).body(response);
-        }
-        return new ResponseEntity<List<Noticia>>(noticia, HttpStatus.OK);
-}
+    public ResponseEntity<?> getAllNoticias() {
+        List<Noticia> noticias = this.noticiaService.listar();
+        return new ResponseEntity<List<Noticia>>(noticias, HttpStatus.OK);
+    }
 
-// Pueba de commit*
-    //:)
-
+    // Listar historial de noticias (para administrador, muestra todas las noticias)
+    @GetMapping("/noticia/historial")
+    public ResponseEntity<?> getHistorialNoticias() {
+        List<Noticia> noticias = this.noticiaService.listarTodas();  // Incluye eliminadas
+        return new ResponseEntity<List<Noticia>>(noticias, HttpStatus.OK);
+    }
 
 }
